refactor(relay): テスト用ソフトウェアをtests/harness/配下に集約
シナリオ本体（data_plane.rs）とテストを支えるソフトウェア（harness）を
ディレクトリで分離。mocksとfixturesはharnessの内部モジュールとなり、
シナリオからはharness::{RelayHarness, ...}だけが見える。
