+++ relay/src/modules/relay/tests.rs
@@ -1,4 +1,2 @@
 mod data_plane;
-mod fixtures;
 mod harness;
-mod mocks;
